+++ packages/ui-kit/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @it-incubator/ui-kit
 
+## 1.0.0
+### Major Changes
+
+- BREAKING: new icon names, consult figma for new names.
+
+- Button now accepts 'as' prop to support links
 ## 0.2.25
 ### Patch Changes